+++ app/subcommands/repo/entrypoint/stop.sh
@@ -1,0 +1,17 @@
+#!/bin/sh
+# Description: does things
+# vim: ft=sh ts=4 sw=4 sts=4 noet
+set -euf
+
+# shellcheck disable=SC1091
+. ./lib.sh
+
+[ -n "${1:-}" ] || fatality 'must provide a repo name'
+
+repo="$1"
+dab repo clone "$repo"
+
+entrypoint="$(config_get "repo/$repo/entrypoint/stop/command")"
+[ -n "$entrypoint" ] || fatality "$repo has no stop entrypoint defined"
+cd "$DAB_REPO_PATH/$repo"
+$entrypoint